export mtl
this edit is not finished. I just got sick of trying to understand how these file formats work and ther is no "Save as draft" like buttin, so I guess this is public now.
obj files refer to ther mtl materials. If ther is a variable for material's names, pleas tell me what it is so I can use that instead of {m_MaterialIDs}_{i}.